Silence dialyzer underspec warnings if possible
In Erlang 24, a -dialyzer({no_underspecs, fn/1}). was added, to silence
dialyzer warnings about the spec being more allowing than the success
typing. It can be difficult to generate very accurate specs.
This is related to the dialyzer -Wunderspec option.

However emit it only if Elrang >= 24, since on ealier Erlangs, it will
result in a warning about an unknown dialyzer warning option.