Rename export-from-git to mk-versioned-archive
Make it possible to run this even if it is not a git archive.
This opens for using it with Github's automatic "Source code" zip/tar
archives, if one sets the version number properly.

Add an option to attempt to pick up the version number from the path
since Github's automatic source code archives can contain gpb-<x.y.z>.

Rename the script since it is now possible to set the version
also when not in a git archive; export-from-git no longer
covers all use cases.

Add a symlink from the old name, export-from-git, for backwards
compatibility, in case someone would have a workflow depending on it.