Allow setting the version in a gpb.vsn file
If a gpb.vsn file exists in the top gpb directory,
its content (the first line) will be used verbatim as version
number.

The concept of version source is added: it is either "git" or "file"
depending on where the version number was found/fetched.

The version source is also written to generated code.

This makes it possible to simplify the mk-versioned-archive script:
it can now just create a gpb.vsn file in the archive.