Remove the old scanner
There have not been any major reported issues with the
new scanner and parser for about a year since its introduction,
so make it the only one.

Also, the old scanner and parser is beginning to lack
in functionality vs the new, risking becoming incompatible,
so it makes no sense any more to keep it around.