Squelch dialyzer warns for translated type encoders
When a (scalar) type always operate on data from a translation
function, then emit a dialyzer nowarn_function attribyte (if
supported for the target erlang version).

The reasoning is that the type encoders are written for the general
case for inputs over the entire supported range, but when all fields
get their data via a translation, it can be that dialyzer can see
that a general-case function clause is dead code with this a
particular translator.