Parse a descriptor structure to defs
The gpb_parse_descr:defs_from_descriptors/1,2
can read a descriptors binary or a #'FileDescriptorSet'{}
record and convert it to definitions, as if it had been
parsed from a .proto file.

The definitions can be used with gpb_compile to generate
encoders and decoders, or it can be used with gpb to encode
and decode messages.