Fix json encoding with mapfields_as_maps option
The json encoder mistakenly selected 2tuples-or-maps based
whether msgs were records or maps, not based on the
setting for mapfields.  The maps option expands to the
corresponding setting for mapfields, but this can also be
set or overridden individually.

The json decoding and the nif functions did the right thing.