Add compiler, crypto and eunit to Dialyzer's plt for OTP 26
In Erlang 26, dialyzer considers it an error when there are calls to
unknown functions, or unknown types, so add the compiler to the plt,
and add crypto as well, because the compiler needs crypto.  Some of
the modules in src have additional eunit tests for internal
functions, so we need to add eunit too.