Expect maps unconditionally, drop Erlang 17
Remove preprocessor ifdef disabling of maps tests.  This means we
drop support for Erlang 17 or earlier.  Maps were introduced in
Erlang 17 already, but Erlang 18 added the possibility for using
variables for map keys, and we need this for flat oneof.

Before this commit, it actually worked all the way back to R14.