Remove output flushing-at-halt workaround
In early Erlang R15 versions as well as in prior versions,
it was necessary to delay for a while before halting, as
an attempt to await all pending text to stdout/stderr getting
printed.

Since we now support Erlang 19 and later, there is no longer
any need for such a workaorund.