Append to ERLC_FLAGS with override for autodetects
Use

  override ERLC_FLAGS += -D<...>

instead of this (without override)

  ERLC_FLAGS += -D<...>

...but do it only for features that are auto-detected from what the
current Erlang supports. These defintions are often used to indicate
that some feature is missing, for example when using an older
version of Erlang.

The "override" makes it possible to specify additional ERLC_FLAGS on
the command line and still get any auto-detected definitions, like
this:

   make ERLC_FLAGS=-Dmy_flag=x

and (on an older Erlang) it might then compile with:

   erlc -Dmy_flag=x -DNO_HAVE_ERL20_STR_FUNCTIONS=true gpb_some_file.erl