Use Erlang 20 string functions, drop Erlang 19
In Erlang 20 (and 19) new functions were added to the string and
list modules. The old functions (in the string module), that were
replacing, only handled latin1, if I remember right. The old
functions got deprecated in Erlang 21.

In gpb, we were previously using the old functions when gpb was
compiled on Erlang 19 or 20. Now unconditionally use the new
functions, thus drop support for Erlang 19 or earlier.