Make the snake_case option idempotent
Make the {rename,What,snake_case} option idempotent, that is the
result after renaming will be the same regardless of how many times
the snake casing option is applied.

For CamelCased names, there is no change.

For some names, this commit introduces non-backwards compatible
changes. The table below attempts to illustrate the behavior:

      Name         snake_case    old behavior    comment

   1  AbcDef       abc_def       abc_def         same
   2  abc_def      abc_def       abc_def         same
   3  Abc_Def      abc_def       abc__def        not same
   4  X097Def      x_097_def     x_097_def       same
   5  x_097_def    x_097_def     x__097_def      not same

Thus, this commit can introduce non-backwards compatible changes for
snake casing of names names...
* that containing capitalized words and underscores
* that containing underscores and digits
* if applied more than once, for example if a framework
  automatically adds it

This change originated from a perceived odd snake casing that was in
fact the result of the snake_case option accidentally applied twice,
like on table rows 4 and 5 above.

Add a renaming option, {rename,_,old_snake_case} to get the old
behaviour in case it would not be possible to adapt for some reason.